Реализован выбор очередности муз. сервисов в Runtime через пропертиз.
