+++ src/main/resources/application.properties
@@ -43,9 +43,9 @@ music.path=music/
 music.initPath=music1/
 
 ## Music Search Services:
-music.searchService.one=krolikSaitServiceImpl
+music.searchService.one=muzofondfmMusicSearchImpl
 music.searchService.two=zaycevSaitServiceImpl
-music.searchService.three=muzofondfmMusicSearchImpl
+music.searchService.three=krolikSaitServiceImpl
 music.searchService.four=downloadMusicVkRuServiceImpl
 
 ## Compilation's covers (pics)
